pyproject.toml
[tool.poetry]
name = "athena_stocks"
version = "0.1.0"
description = "One Stop Web App for Data Analytics tool for stock trading"
authors = ["VinodhKotipalli <50426917+VinodhKotipalli@users.noreply.github.com>"]
license = "MIT Licence"

[tool.poetry.dependencies]
python = "^3.7.1"
numpy = "^1.21.0"
pandas = "^1.3.0"
selenium = "^3.141.0"
bs4 = "^0.0.1"
newspaper3k = "^0.2.8"
Django = "^3.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
